Create Day_3-Strings-Only
This function works the same as the other function in that the user is prompted to enter a word. The function counts the number of vowels in the word and displays the word entered and the number of vowels. 
The difference is that this function will not allow integers be entered.